+++ R/aml_score_data.R
@@ -3,6 +3,8 @@
 #' @description  Function is using the latest data from the financial assets indicator pattern and deep learning model.
 #' Prediction is a price change in the future for that asset will be used by the trading system
 #'
+#' `r lifecycle::badge('stable')`
+#'
 #' @details Performs fresh data reading from the rds file
 #'
 #'