+++ R/datasets.R
@@ -109,18 +109,6 @@
 #' }
 "price_dataset_big"
 
-#' Table with indicator dataset, 30000 rows
-#'
-#' @docType data
-#' @keywords datasets
-#'
-#' @format A dataframe with several columns
-#' \describe{
-#'   \item{X1}{Date and time of the indicator sample}
-#'   \item{X2-X29}{Values of the assets}
-#' }
-"indicator_dataset_big"
-
 #' Table with one column indicator dataset
 #'
 #' @docType data
@@ -198,7 +186,7 @@
 #' }
 "x_test_model"
 
-#' Table with predicte price change
+#' Table with predicted price change
 #'
 #' @docType data
 #' @keywords datasets