+++ R/mt_stat_evaluate.R
@@ -4,6 +4,7 @@
 #'                       to find out specific market type of the financial asset
 #'                       it will also discard bad result outputting -1 if it is the case
 #'
+#' `r lifecycle::badge('stable')`
 #'
 #' @details it is mandatory to switch on the virtual h2o machine with h2o.init()
 #' also to shut it down with h2o.shutdown(prompt = F)