+++ R/rl_log_progress.R
@@ -3,6 +3,7 @@
 #' @description Function will record Q values during the model update. These values will be used by another function
 #' Function was developed to help to estimate best control parameters during optimisation process
 #'
+#' `r lifecycle::badge('stable')`
 #'
 #' @param x - dataframe containing trading results
 #' @param states  - Selected states of the System