+++ R/rl_record_policy_mt.R
@@ -2,6 +2,10 @@
 #'
 #' @description Function will write a policy 'decision' to the csv file specific for each Expert Advisor
 #'
+#' `r lifecycle::badge('stable')`
+#'
+#' @details It is imperative that terminal path contains exact word Terminal3
+#'
 #' @param x - Dataframe containing columns MarketType and Policy
 #' @param trading_system - numeric vector of length 1 with Trading System Magic Number information
 #' @param path_terminal - string, path to the terminal where this Policy/Decision must be written
@@ -31,7 +35,7 @@ rl_record_policy_mt <- function(x, trading_system, path_terminal, fileName = "Sy
 # derive which terminal should be enabled (using path to sandbox) and using variable 'addition'
   is_T3 <- stringr::str_detect(path_terminal, "Terminal3")
   is_T4 <- stringr::str_detect(path_terminal, "Terminal4")
-  if(is_T3 == TRUE) { addition <- 200 } else if(is_T4 == TRUE) { addition <- 300 } else { addition <- 0 }
+  if(is_T3 == TRUE) { addition <- 200 } else if(is_T4 == TRUE) { addition <- 300 } else { addition <- 200 }
 
   # -------------------------
   # Write Decision/Update Policy